# docs/phase-1/mango-mdu-openapi.yaml
openapi: 3.0.3
info:
  title: Mango MDU Service API
  version: 1.0.0
  description: |
    Northbound API exposed by MDU for Phase 1 workflows.

    This contract defines the API consumed by northbound clients such as the MDU UI.
    It intentionally normalizes downstream platform behavior into stable MDU-owned resources.

    Phase 1 integrations:
    - OWSEC: authentication, token validation, and user identity/user CRUD handled directly outside MDU
    - PROV: entities, venues, effective scope resolution

    Phase 1 scope:
    - session/access context, hierarchy/entity/venue wrappers, and user-access orchestration only

    Explicit non-goal:
    - billing features are out-of-scope for Phase 1
servers:
  - url: /api/v1
    description: Relative API base path
...
security:
  - bearerAuth: []
paths:
...
  /entities:
    get:
      tags: [Entities]
      summary: List entities visible to the caller
      description: Returns normalized entity resources filtered to the caller's effective scope.
      operationId: listEntities
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XCorrelationIdHeader'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Entity list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityListResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '503':
          $ref: '#/components/responses/BackendUnavailableError'
    post:
      tags: [Entities]
      summary: Create an entity
      description: Creates a new Phase 1 entity resource.
      operationId: createEntity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEntityRequest'
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XCorrelationIdHeader'
      responses:
        '201':
          description: Entity created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityDetail'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '503':
          $ref: '#/components/responses/BackendUnavailableError'
  /entities/{entityId}:
    get:
      tags: [Entities]
      summary: Get entity detail
      operationId: getEntity
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XCorrelationIdHeader'
        - $ref: '#/components/parameters/EntityIdPath'
      responses:
        '200':
          description: Entity detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityDetail'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '503':
          $ref: '#/components/responses/BackendUnavailableError'
    put:
      tags: [Entities]
      summary: Update an entity
      operationId: updateEntity
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XCorrelationIdHeader'
        - $ref: '#/components/parameters/EntityIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEntityRequest'
      responses:
        '200':
          description: Entity updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityDetail'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '503':
          $ref: '#/components/responses/BackendUnavailableError'
    delete:
      tags: [Entities]
      summary: Delete an entity
      operationId: deleteEntity
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XCorrelationIdHeader'
        - $ref: '#/components/parameters/EntityIdPath'
      responses:
        '204':
          description: Entity deleted
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '503':
          $ref: '#/components/responses/BackendUnavailableError'
...
  /users/{userId}/assignments:
    get:
      tags: [Users]
      summary: Get effective scoped assignments for a user
      description: Returns the target user's effective scoped assignments as normalized MDU resources.
      operationId: getUserAssignments
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XCorrelationIdHeader'
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: User scope assignments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAssignmentsResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '503':
          $ref: '#/components/responses/BackendUnavailableError'
    post:
      tags: [Users]
      summary: Assign a user to an entity or venue scope
      description: |
        Assigns the target user to the specified entity or venue scope with a given role.
        
        This endpoint implements an idempotent resolution flow when dealing with existing states:
        1. **Create (201 Created)**: If no matching scoped role or assignment exists for the requested scope and role, MDU creates the assignment.
        2. **Update/Resolve (201 Created)**: If a matching scoped role exists downstream for that scope and role, but the target user is not yet bound to it, MDU updates/resolves the role by adding the target user to its user list.
        3. **No-Op/Idempotent Success (200 OK)**: If the target user is already assigned to the exact scope and role combination, the request resolves immediately with no state changes.
        4. **Conflict (409 Conflict)**: Reserved only for true, unresolvable state conflicts (e.g. if the requested assignment cannot be safely resolved from downstream state or creates a logical contradiction).
      operationId: assignUserToScope
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XCorrelationIdHeader'
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserAssignmentRequest'
      responses:
        '200':
          description: Requested assignment already exists and is active (no-op/idempotent success).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAssignment'
              examples:
                alreadyAssigned:
                  summary: Already assigned (idempotent success)
                  value:
                    assignmentId: "523e4567-e89b-12d3-a456-426614174000"
                    scopeType: "entity"
                    scopeId: "223e4567-e89b-12d3-a456-426614174000"
                    scopeName: "Regional Operations"
                    role: "admin"
                    path:
                      - id: "223e4567-e89b-12d3-a456-426614174000"
                        name: "Regional Operations"
                        type: "entity"
                    managementRoleId: "623e4567-e89b-12d3-a456-426614174000"
                    managementPolicyId: "523e4567-e89b-12d3-a456-426614174000"
                    createdAt: "2026-06-30T12:00:00Z"
        '201':
          description: Assignment newly created, or existing matching role updated to newly add the user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAssignment'
              examples:
                newAssignment:
                  summary: New assignment created
                  value:
                    assignmentId: "523e4567-e89b-12d3-a456-426614174000"
                    scopeType: "entity"
                    scopeId: "223e4567-e89b-12d3-a456-426614174000"
                    scopeName: "Regional Operations"
                    role: "admin"
                    path:
                      - id: "223e4567-e89b-12d3-a456-426614174000"
                        name: "Regional Operations"
                        type: "entity"
                    managementRoleId: "623e4567-e89b-12d3-a456-426614174000"
                    managementPolicyId: "523e4567-e89b-12d3-a456-426614174000"
                    createdAt: "2026-06-30T12:00:00Z"
                existingRoleUpdated:
                  summary: Existing matching role updated to add user
                  value:
                    assignmentId: "723e4567-e89b-12d3-a456-426614174000"
                    scopeType: "venue"
                    scopeId: "323e4567-e89b-12d3-a456-426614174000"
                    scopeName: "East Wing Venue"
                    role: "installer"
                    path:
                      - id: "223e4567-e89b-12d3-a456-426614174000"
                        name: "Regional Operations"
                        type: "entity"
                      - id: "323e4567-e89b-12d3-a456-426614174000"
                        name: "East Wing Venue"
                        type: "venue"
                    managementRoleId: "823e4567-e89b-12d3-a456-426614174000"
                    managementPolicyId: "723e4567-e89b-12d3-a456-426614174000"
                    createdAt: "2026-06-30T12:00:00Z"
        '400':
          description: Invalid request parameters or body (e.g. invalid scopeType or missing required fields).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                invalidScopeType:
                  summary: Invalid scope type value
                  value:
                    ErrorCode: 400
                    ErrorDetails: "The field 'scopeType' must be either 'entity' or 'venue'."
                    ErrorDescription: "Bad Request"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          description: Conflict (unresolvable state conflicts).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                unresolvedConflict:
                  summary: True state conflict
                  value:
                    ErrorCode: 409
                    ErrorDetails: "Cannot safely resolve requested assignment from existing downstream state."
                    ErrorDescription: "Conflict"
        '503':
          $ref: '#/components/responses/BackendUnavailableError'
  /users/{userId}/assignments/{assignmentId}:
    delete:
      tags: [Users]
      summary: Remove a user's scope assignment
      description: Removes the scoped access binding for the target user by assignment ID.
      operationId: deleteUserAssignment
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XCorrelationIdHeader'
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/AssignmentIdPath'
      responses:
        '204':
          description: Assignment removed
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '503':
          $ref: '#/components/responses/BackendUnavailableError'
  /users/{userId}/access-policy:
    get:
      tags: [Users]
      summary: Retrieve the access policy for a user at a given scope
      description: Fetches the configured management policy and template role for the target user at the specified entity or venue scope.
      operationId: getUserAccessPolicy
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XCorrelationIdHeader'
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/ScopeQueryRequired'
        - $ref: '#/components/parameters/EntityIdQueryRequired'
        - $ref: '#/components/parameters/VenueIdQuery'
      responses:
        '200':
          description: Access policy details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccessPolicy'
        '400':
          description: Invalid query parameters (e.g. missing entityId, or venue scope missing venueId).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                missingEntityId:
                  summary: Missing entityId parameter
                  value:
                    ErrorCode: 400
                    ErrorDetails: "The required query parameter 'entityId' is missing."
                    ErrorDescription: "Bad Request"
                missingVenueId:
                  summary: Venue scope missing venueId
                  value:
                    ErrorCode: 400
                    ErrorDetails: "The query parameter 'venueId' is required when scope is set to 'venue'."
                    ErrorDescription: "Bad Request"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '503':
          $ref: '#/components/responses/BackendUnavailableError'
    put:
      tags: [Users]
      summary: Assign or update the access policy for a user
      description: Sets or updates the management role, policies, and resource permissions for the target user at the specified entity or venue scope.
      operationId: updateUserAccessPolicy
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XCorrelationIdHeader'
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAccessPolicy'
      responses:
        '200':
          description: Access policy updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccessPolicy'
        '400':
          description: Invalid payload combination (e.g. scope is entity but venueId is provided, or scope is venue but venueId is missing).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                invalidScopePayload:
                  summary: Invalid scope payload
                  value:
                    ErrorCode: 400
                    ErrorDetails: "venueId must not be provided when scope is 'entity'."
                    ErrorDescription: "Bad Request"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Targeted user or scoped role/policy binding cannot be resolved downstream.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                bindingNotFound:
                  summary: User or access policy bindings not found
                  value:
                    ErrorCode: 404
                    ErrorDetails: "The access policy configuration for the target user at the specified scope does not exist."
                    ErrorDescription: "Not Found"
        '409':
          $ref: '#/components/responses/ConflictError'
        '503':
          $ref: '#/components/responses/BackendUnavailableError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OWSEC-issued user bearer token authentication. Note that token validation is conditionally enforced on the public interface and can be bypassed for local development and testing by setting the environment variable AUTH_ENABLED=false.
  parameters:
    XRequestIdHeader:
      name: X-Request-Id
      in: header
      required: false
      description: Unique tracking identifier for this specific HTTP request. If not supplied, the service will generate one.
      schema:
        type: string
    XCorrelationIdHeader:
      name: X-Correlation-Id
      in: header
      required: false
      description: Distributed transaction identifier linking requests across multiple microservices. If not supplied, the service will fall back to using X-Request-Id (or generate one) to preserve trace context.
      schema:
        type: string
    EntityIdPath:
      name: entityId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Id'
...
    UserIdPath:
      name: userId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Id'
...
    EntityIdQueryRequired:
      name: entityId
      in: query
      required: true
      description: Target entity scope. Always required. If scope is venue, both entityId and venueId must be specified.
      schema:
        $ref: '#/components/schemas/Id'
    VenueIdQuery:
      name: venueId
      in: query
      required: false
      description: Target venue scope. Required only when scope is venue (in which case both entityId and venueId must be specified).
      schema:
        $ref: '#/components/schemas/Id'
    AssignmentIdPath:
      name: assignmentId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Id'
...
    ScopeQueryRequired:
      name: scope
      in: query
      required: true
      description: The target access policy scope (entity or venue).
      schema:
        type: string
        enum: [entity, venue]
    LimitQuery:
      name: limit
      in: query
      required: false
      description: Maximum number of items to return.
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    OffsetQuery:
      name: offset
      in: query
      required: false
      description: Number of items to skip.
      schema:
        type: integer
        default: 0
        minimum: 0
...
  responses:
    BadRequestError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            badRequestDefault:
              summary: Validation error default
              value:
                ErrorCode: 400
                ErrorDetails: "One or more request parameters are invalid."
                ErrorDescription: "Bad Request"
    UnauthorizedError:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            unauthorizedDefault:
              summary: Unauthorized access error
              value:
                ErrorCode: 401
                ErrorDetails: "Bearer token is missing, expired, or invalid."
                ErrorDescription: "Unauthorized"
    ForbiddenError:
      description: Caller is not allowed to perform this action in the resolved scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            forbiddenDefault:
              summary: Forbidden access error
              value:
                ErrorCode: 403
                ErrorDetails: "You do not have permission to access resources in the resolved scope."
                ErrorDescription: "Forbidden"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            notFoundDefault:
              summary: Resource not found error
              value:
                ErrorCode: 404
                ErrorDetails: "The requested resource could not be found."
                ErrorDescription: "Not Found"
    ConflictError:
      description: Current state conflicts with the requested operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            conflictDefault:
              summary: Conflict error default
              value:
                ErrorCode: 409
                ErrorDetails: "The operation conflicts with the current state of the resource."
                ErrorDescription: "Conflict"
    BackendUnavailableError:
      description: One or more required downstream services are unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            backendUnavailableDefault:
              summary: Backend service unavailable
              value:
                ErrorCode: 503
                ErrorDetails: "A required downstream service (e.g. PROV or OWSEC) is currently unreachable or timed out."
                ErrorDescription: "Service Unavailable"
  schemas:
    Id:
      type: string
      description: UUID or service-owned opaque identifier
    IsoDateTime:
      type: string
      format: date-time
    ListMetadata:
      type: object
      required: [total, limit, offset]
      properties:
        total:
          type: integer
          description: Total count of items matching the filter.
        limit:
          type: integer
          description: Maximum number of items requested.
        offset:
          type: integer
          description: Offset of the returned items.
    HierarchyNodeType:
      type: string
      enum: [operator, entity, venue]
    AssignmentScopeType:
      type: string
      enum: [entity, venue]
    RoleKey:
      type: string
      enum: [root, admin, csr, installer, noc, accounting, system]
    ApiError:
      type: object
      required: [ErrorCode, ErrorDetails, ErrorDescription]
      properties:
        ErrorCode:
          type: integer
          description: Numeric error code (typically matching the HTTP status code or specific REST sub-code).
        ErrorDetails:
          type: string
          description: Detailed explanation or context of the error.
        ErrorDescription:
          type: string
          description: High-level classification of the error category.
      example:
        ErrorCode: 400
        ErrorDetails: "The requested field 'email' is invalid."
        ErrorDescription: "Bad Request"
    ScopePathItem:
      type: object
      required: [id, type, name]
      properties:
        id:
          $ref: '#/components/schemas/Id'
        type:
          $ref: '#/components/schemas/HierarchyNodeType'
        name:
          type: string
...
    EntitySummary:
      type: object
      required: [id, name, type, path]
      properties:
        id:
          $ref: '#/components/schemas/Id'
        name:
          type: string
        parentId:
          allOf:
            - $ref: '#/components/schemas/Id'
          nullable: true
        type:
          type: string
          enum: [normal, subscriber]
        path:
          type: array
          items:
            $ref: '#/components/schemas/ScopePathItem'
        venueCount:
          type: integer
        userCount:
          type: integer
        deviceCount:
          type: integer
    EntityDetail:
      allOf:
        - $ref: '#/components/schemas/EntitySummary'
        - type: object
          properties:
            description:
              type: string
            operatorId:
              $ref: '#/components/schemas/Id'
            managementPolicyId:
              $ref: '#/components/schemas/Id'
            managementRoleIds:
              type: array
              items:
                $ref: '#/components/schemas/Id'
            createdAt:
              $ref: '#/components/schemas/IsoDateTime'
            updatedAt:
              $ref: '#/components/schemas/IsoDateTime'
      example:
        id: "223e4567-e89b-12d3-a456-426614174000"
        name: "Regional Operations"
        parentId: "123e4567-e89b-12d3-a456-426614174000"
        type: "normal"
        path:
          - id: "223e4567-e89b-12d3-a456-426614174000"
            name: "Regional Operations"
            type: "entity"
        venueCount: 12
        userCount: 5
        deviceCount: 45
        description: "Regional office entity"
        operatorId: "323e4567-e89b-12d3-a456-426614174000"
        managementPolicyId: "523e4567-e89b-12d3-a456-426614174000"
        managementRoleIds:
          - "623e4567-e89b-12d3-a456-426614174000"
        createdAt: "2026-06-30T12:00:00Z"
        updatedAt: "2026-06-30T12:00:00Z"
    EntityListResponse:
      type: object
      required: [items, metadata]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EntitySummary'
        metadata:
          $ref: '#/components/schemas/ListMetadata'
    CreateEntityRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        parentEntityId:
          $ref: '#/components/schemas/Id'
        type:
          type: string
          enum: [normal, subscriber]
          default: normal
    UpdateEntityRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...
    UserAssignment:
      type: object
      required: [assignmentId, scopeType, scopeId, scopeName, role, path]
      properties:
        assignmentId:
          $ref: '#/components/schemas/Id'
        scopeType:
          $ref: '#/components/schemas/AssignmentScopeType'
        scopeId:
          $ref: '#/components/schemas/Id'
        scopeName:
          type: string
        role:
          $ref: '#/components/schemas/RoleKey'
        path:
          type: array
          items:
            $ref: '#/components/schemas/ScopePathItem'
        managementRoleId:
          $ref: '#/components/schemas/Id'
        managementPolicyId:
          $ref: '#/components/schemas/Id'
        createdAt:
          $ref: '#/components/schemas/IsoDateTime'
      example:
        assignmentId: "523e4567-e89b-12d3-a456-426614174000"
        scopeType: "entity"
        scopeId: "223e4567-e89b-12d3-a456-426614174000"
        scopeName: "Regional Operations"
        role: "admin"
        path:
          - id: "223e4567-e89b-12d3-a456-426614174000"
            name: "Regional Operations"
            type: "entity"
        managementRoleId: "623e4567-e89b-12d3-a456-426614174000"
        managementPolicyId: "523e4567-e89b-12d3-a456-426614174000"
        createdAt: "2026-06-30T12:00:00Z"
    UserAssignmentsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserAssignment'
    CreateUserAssignmentRequest:
      type: object
      required: [scopeType, scopeId, role]
      properties:
        scopeType:
          $ref: '#/components/schemas/AssignmentScopeType'
        scopeId:
          $ref: '#/components/schemas/Id'
        role:
          $ref: '#/components/schemas/RoleKey'
...
    UserAccessPolicy:
      oneOf:
        - $ref: '#/components/schemas/EntityUserAccessPolicy'
        - $ref: '#/components/schemas/VenueUserAccessPolicy'
...